Consolidated dataset configs, tested and debugged sliding window through plot generation
