+++ SLI_pipeline/README.md
@@ -1,9 +1,15 @@
 # Sea Level Indicators Pipeline
 
-## Harvests
+## run_pipeline.py
 
-## Aggregates data into cycles
+## harvester.py
 
-## Grids along track data
+## cycle_gridding.py
 
-## Calculates indicator values
+## indicators.py
+
+## txt_engine.py
+
+## plotting/plot_generation.py
+
+## upload_indicators.py
