+++ SLI_pipeline/configs/datasets.yaml
@@ -1,0 +1,45 @@
+- ds_name: "ERS_1"
+  harvester_type: "PODAAC Drive"
+  start: "19920920T00:00:00Z"
+  end: "19950424T23:59:59Z"
+  most_recent: False
+- ds_name: "ERS_2"
+  harvester_type: "PODAAC Drive"
+  start: "19950425T00:00:00Z"
+  end: "20020523T23:59:59Z"
+  most_recent: False
+- ds_name: "ENVISAT_1"
+  harvester_type: "PODAAC Drive"
+  start: "20020524T00:00:00Z"
+  end: "20111231T23:59:59Z"
+  most_recent: False
+- ds_name: "CYROSAT_2"
+  harvester_type: "PODAAC Drive"
+  start: "20120101T00:00:00Z"
+  end: "20130314T23:59:59Z"
+  most_recent: False
+- ds_name: "SARAL"
+  harvester_type: "PODAAC Drive"
+  start: "20130315T00:00:00Z"
+  end: "20160223T23:59:59Z"
+  most_recent: False
+- ds_name: "SENTINEL_3A"
+  harvester_type: "PODAAC Drive"
+  start: "20160224T00:00:00Z"
+  end: "20180628T23:59:59Z"
+  most_recent: False
+- ds_name: "SENTINEL_3B"
+  harvester_type: "PODAAC Drive"
+  start: "20180629T00:00:00Z"
+  end: "20220101T23:59:59Z"
+  most_recent: True
+- ds_name: "MERGED_ALT"
+  harvester_type: "PODAAC Drive"
+  start: "19920925T00:00:00Z"
+  end: "20220101T23:59:59Z"
+  most_recent: True
+- ds_name: "JASON_3"
+  harvester_type: "PODAAC Drive"
+  start: "20210401T00:00:00Z"
+  end: "20220101T23:59:59Z"
+  most_recent: True
