+++ SLI_pipeline/configs/login.yaml.example
