--- SLI_pipeline/dataset_configs/SENTINEL_3B.yaml
@@ -1,25 +1,0 @@
-# =====================================================
-# Harvester
-# =====================================================
-harvester_type: "PODAAC Drive"
-start: "20180629T00:00:00Z" # yyyymmddThh:mm:ssZ
-end: "20181231T23:59:59Z" # yyyymmddThh:mm:ssZ
-most_recent: True # Overrides end date to be current date at time of execution
-date_regex: "%Y-%m-%dT%H:%M:%SZ" # does not change
-
-# =====================================================
-# Dataset
-# =====================================================
-ds_name: "SENTINEL_3B" # Name for dataset
-
-# new_data_attrs:
-original_dataset_title: SENTINEL 3B
-original_dataset_short_name: SENTINEL 3B
-original_dataset_url: "N/A"
-original_dataset_reference: "N/A"
-original_dataset_doi: "Not Registered"
-
-# Processing:
-version: 1.0
-cycle_length: 30
-data_type: "along track"
